+++ datafusion/catalog/Cargo.toml
@@ -28,6 +28,8 @@ rust-version.workspace = true
 version.workspace = true
 
 [dependencies]
+arrow = { workspace = true }
+arrow-array = { workspace = true }
 arrow-schema = { workspace = true }
 async-trait = { workspace = true }
 dashmap = { workspace = true }
@@ -35,8 +37,12 @@ datafusion-common = { workspace = true }
 datafusion-execution = { workspace = true }
 datafusion-expr = { workspace = true }
 datafusion-physical-plan = { workspace = true }
+datafusion-sql = { workspace = true }
+futures = { workspace = true }
 itertools = { workspace = true }
+log = { workspace = true }
 parking_lot = { workspace = true }
+sqlparser = { workspace = true }
 
 [dev-dependencies]
 tokio = { workspace = true }
