+++ datafusion/core/src/datasource/mod.rs
@@ -33,12 +33,9 @@ pub mod provider;
 pub mod schema_adapter;
 mod statistics;
 pub mod stream;
-pub mod streaming;
 pub mod view;
 
 // backwards compatibility
-pub use datafusion_execution::object_store;
-
 pub use self::default_table_source::{
     provider_as_source, source_as_provider, DefaultTableSource,
 };
@@ -46,6 +43,7 @@ pub use self::memory::MemTable;
 pub use self::view::ViewTable;
 pub use crate::catalog::TableProvider;
 pub use crate::logical_expr::TableType;
+pub use datafusion_execution::object_store;
 pub use statistics::get_statistics_with_limit;
 
 use arrow_schema::{Schema, SortOptions};
