+++ datafusion/substrait/tests/utils.rs
@@ -17,7 +17,7 @@
 
 #[cfg(test)]
 pub mod test {
-    use datafusion::catalog_common::TableReference;
+    use datafusion_catalog::TableReference;
     use datafusion::common::{substrait_datafusion_err, substrait_err};
     use datafusion::datasource::empty::EmptyTable;
     use datafusion::datasource::TableProvider;
